% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_identifier_line_to_classification.R
\name{add_identifier_line_to_classification}
\alias{add_identifier_line_to_classification}
\title{Adds an identifier line at the bottom of a classification dataframe created by taxize::classification or obtained from BOLD}
\usage{
add_identifier_line_to_classification(classif, identifier_value,
  classification_source = "ncbi")
}
\arguments{
\item{classif}{classification data frame for an individual taxon, created by taxize::classification}

\item{identifier_value}{string containing value of the identifier. Good idea to specify as a key:value pair (e.g. 'uid:1109490022')}

\item{classification_source}{string specifying the source of the classification information: either "ncbi" or "bold" is allowed.}
}
\description{
Adds an identifier line at the bottom of a classification dataframe created by taxize::classification or obtained from BOLD
}
\examples{
add_identifier_line_to_classification(classif, identifier_value='uid:1109490022')
}
\keyword{classification}
\keyword{dada2}
\keyword{ncbi}
\keyword{taxize}
\keyword{taxonomy}
\keyword{uid}
