% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_do_classification.R
\name{check_and_do_classification}
\alias{check_and_do_classification}
\title{check_and_do_classification}
\usage{
check_and_do_classification(taxon, classification_file_name, id_list,
  chunk_size = 60, block_size = 5000)
}
\arguments{
\item{classification_file_name}{Path to the classification file (character string)}

\item{id_list}{A list of character vectors of NCBI UIDs, such as that produced by extract_ids_from_ncbi_search_blocks.}

\item{chunk_size}{An integer giving the size of chunks into which each id vector should be split. This is needed to return the results properly}

\item{block_size}{An integer giving the size of blocks into which the full vector of IDs with missing classification info is split. Default = 5000}
}
\description{
This function performs taxonomic classification using the taxize::classification function,
on a list of vectors of NCBI UIDs such as that returned by extract_ids_from_ncbi_search_blocks.
This function is quite slow because it is unable to use the web history (this produces only unique taxon IDs, which will not map 1-to-1 to sequence UIDs)
It returns the extracted classification values, one per id, in a list.
}
\examples{
check_and_do_classification(taxon="Hypogasturidae", classification_file_name=make_classification_file_name(taxon="Hypogasturidae", directory="~/Documents/classification_files"), id_list=list('1072897717', '329757399', '259881131'),  chunk_size=60)
}
\keyword{classification}
\keyword{ncbi}
\keyword{taxonomy}
\keyword{uid}
