% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_do_mapping.R
\name{check_and_do_mapping}
\alias{check_and_do_mapping}
\title{Check for and Perform ID Mapping}
\usage{
check_and_do_mapping(taxon, mapping_file_name, id_list, chunk_size = 60)
}
\arguments{
\item{taxon}{Taxon name as a character string.}

\item{mapping_file_name}{Path to the mapping file}

\item{chunk_size}{An integer giving the size of chunks into which id_vector will be split for efficient extraction. Defaults to 60.}

\item{id_vector}{A character vector of NCBI UIDs}
}
\description{
This function checks for an existing mapping file for the relevant taxon
in the given input directory with the given suffix
If the mapping file already exists, it checks whether any NCBI UIDs are missing
and adds the mapping info (NCBI UID, BOLD (Barcode of Life Database) ID and NCBI accession.version ID)
for those UIDs.
If the mapping file doesn't yet exist, it creates the file and performs the ID mapping.
Mapping is done 'chunk-wise' to fit with the NCBI API limitations.
If any errors are encountered during the mapping, (e.g. in connecting to the Entrez API), the relevant chunk will be skipped
and the extraction will continue.
}
\examples{
check_and_do_mapping(taxon="Hypogasturidae", mapping_file_name, id_list <- list(c('1072897717', '329757399', '259881131')), chunk_size=60)
}
\keyword{bold}
\keyword{id}
\keyword{ncbi}
