% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get_bold_records.R
\name{check_and_get_bold_records}
\alias{check_and_get_bold_records}
\title{check_and_get_bold_records}
\usage{
check_and_get_bold_records(taxon, marker, mapping_file_name, dada2_file_name)
}
\arguments{
\item{taxon}{a taxon name as character string}

\item{marker}{One or more marker names in BOLD format (character string with pipe delimiter separating names)}

\item{mapping_file_name}{Path to mapping file created with check_and_do_mapping command}

\item{dada2_file_name}{Path to output file}
}
\description{
This function examines an NCBI mapping file to see which BOLD records were already obtained from NCBI.
It then obtains the remaining public sequences from BOLD for the given taxon and marker combination.
It outputs a fasta file containing marker sequence with the taxonomy information (names only) for 
kingdom, phylum, class, order, family, genus, and species presented in the format required for
a dada2 reference database.
NB THIS FUNCTION CURRENTLY DOESN'T WORK - SEEMS TO BE DUE TO A BUG IN bold::bold_specimens
Have raised at https://github.com/ropensci/bold/issues/46
}
\examples{
check_and_get_bold_records(taxon="Athericidae", marker="COI-5P|COI-3P", mapping_file_name=make_mapping_name(taxon="Athericidae", directory="~/Documents/output/"), dada2_file_name=paste("~/Documents/output/", "Athericidae_extra_bold_seq_dada2.fasta", sep=""))
}
\keyword{bold}
\keyword{dada2}
\keyword{fasta}
\keyword{metabarcoding}
\keyword{ncbi}
