% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_id_vector.R
\name{classify_id_vector}
\alias{classify_id_vector}
\title{Perform Taxonomic Classification on a Vector of NCBI Unique Identifiers}
\usage{
classify_id_vector(id_vector, chunk_size = 60, taxon)
}
\arguments{
\item{id_vector}{A character vector of NCBI UIDs}

\item{chunk_size}{An integer giving the size of chunks into which the id vector is split. This is needed to return the results properly}
}
\description{
This function performs taxonomic classification using the taxize::classification function,
on a vector of NCBI UIDs.
This function is quite slow because it is unable to use the web history (this produces only unique taxon IDs, which will not map 1-to-1 to sequence UIDs)
It returns the extracted classification values, one per id, in a list.
Each classification value is a dataframe (see taxize::classification for more detail.)
}
\examples{
classify_id_vector(id_vector=c('1072897717', '329757399', '259881131'), taxon="Hypogasturidae", chunk_size=60)
}
\keyword{classification}
\keyword{ncbi}
\keyword{taxonomy}
\keyword{uid}
