% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_names_to_ncbi_query_text.R
\name{convert_gene_names_to_ncbi_query_text}
\alias{convert_gene_names_to_ncbi_query_text}
\title{convert_gene_names_to_ncbi_query_text}
\usage{
convert_gene_names_to_ncbi_query_text(gene_names)
}
\arguments{
\item{gene_names}{A single gene name (as character) or a vector of gene names (as character)}
}
\description{
This function converts a single gene name or a vector of alternative gene names
to properly-formatted text to include in a NCBI search query
}
\examples{
convert_gene_names_to_ncbi_query_text(gene_names=c("COI", "CO1", "COX1"))
}
\keyword{gene}
\keyword{ncbi}
\keyword{sequence}
