% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_extraction_functions.R
\name{extract_acc_ver_ids_from_raw_fasta_object}
\alias{extract_acc_ver_ids_from_raw_fasta_object}
\title{extract_acc_ver_ids_from_raw_fasta_object}
\usage{
extract_acc_ver_ids_from_raw_fasta_object(fasta_object)
}
\arguments{
\item{fasta_object}{A FASTA object: a list of class 'DNAbin' as produced by ape::read.FASTA}
}
\description{
This function extracts NCBI accession.version ID from a NCBI-retrieved FASTA object
in the form of a list of class 'DNAbin' (i.e. as produced by ape::read.FASTA)
It returns the extracted IDs only, as a list of character vectors
}
\examples{
extract_acc_ver_ids_from_raw_fasta_object(fasta_object="Chlorophyta.fasta")
}
\keyword{ncbi}
\keyword{taxonomy}
\keyword{uid}
