% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_accn_version_from_entrez_xml.R
\name{extract_accn_version_from_entrez_xml}
\alias{extract_accn_version_from_entrez_xml}
\title{extract_accn_version_from_entrez_xml}
\usage{
extract_accn_version_from_entrez_xml(entrez_xml)
}
\arguments{
\item{entrez_xml}{An xml object returned by the rentrez::entrez_fetch command}
}
\description{
This function extracts the accession.version IDs
from an NCBI xml record returned by the rentrez::entrez_fetch command.
It returns a vector of accession.version IDs
}
\examples{
extract_accn_version_from_entrez_xml(entrez_xml="Chlorophyta_xml")
}
\keyword{entrez}
\keyword{ncbi}
\keyword{xml}
