% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bold_and_accn.R
\name{extract_bold_and_accn}
\alias{extract_bold_and_accn}
\title{Extract NCBI Accession.Version IDS and BOLD IDs for a Vector of NCBI UIDs}
\usage{
extract_bold_and_accn(id_vector)
}
\arguments{
\item{id_vector}{A character vector of NCBI UIDs}
}
\description{
This function extracts the BOLD (Barcode of Life Database) IDs, if present, and the accession.version IDs
from NCBI for a input character vector of NCBI UIDs.
It returns a dataframe with columns uid, bold, and accn.
}
\examples{
extract_bold_and_accn(id_vector=c('1072897717', '329757399', '259881131'))
}
\keyword{bold}
\keyword{id}
\keyword{ncbi}
