% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bold_from_entrez_xml.R
\name{extract_bold_from_entrez_xml}
\alias{extract_bold_from_entrez_xml}
\title{extract_bold_from_entrez_xml}
\usage{
extract_bold_from_entrez_xml(entrez_xml)
}
\arguments{
\item{entrez_xml}{An xml object returned by the rentrez::entrez_fetch command}
}
\description{
This function extracts the BOLD (Barcode of Life Database) IDs, if present,
from an NCBI xml record returned by the rentrez::entrez_fetch command.
It returns a vector of BOLD IDs (or NA if no BOLD ID was present for a given entry)
}
\examples{
extract_bold_from_entrez_xml(entrez_xml="Chlorophyta_xml")
}
\keyword{entrez}
\keyword{ncbi}
\keyword{xml}
