% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_extraction_functions.R
\name{extract_ids_from_classification_file}
\alias{extract_ids_from_classification_file}
\title{extract_ids_from_classification_file}
\usage{
extract_ids_from_classification_file(classification_file_name)
}
\arguments{
\item{classification_file_name}{A classification file (RDS format) created by the function check_and_do_classification}
}
\description{
This function extracts NCBI UIDs from a classification file
created with the check_and_do_classification function
It returns the extracted IDs only, as a character vector
}
\examples{
extract_ids_from_classification_file(classification_file_name="Chlorophyta_classification.rds")
}
\keyword{ncbi}
\keyword{taxonomy}
\keyword{uid}
