% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_extraction_functions.R
\name{extract_ids_from_ncbi_search_blocks}
\alias{extract_ids_from_ncbi_search_blocks}
\title{extract_ids_from_ncbi_search_blocks}
\usage{
extract_ids_from_ncbi_search_blocks(ncbi_search_blocks)
}
\arguments{
\item{ncbi_search_blocks}{An object created with the get_ncbi_search_blocks function}
}
\description{
This function extracts NCBI UIDs from a ncbi_search_blocks object
created with the get_ncbi_search_blocks function.
It returns the extracted IDs only, as a list of lists *** IS THIS RIGHT? LIST OF VECTORS? ***
}
\examples{
extract_ids_from_ncbi_search_blocks(ncbi_search_blocks=Chlorophyta_blocks)
}
\keyword{ncbi}
\keyword{uid}
