% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bold_specimens.R
\name{filter_bold_specimens}
\alias{filter_bold_specimens}
\title{filter_bold_specimens}
\usage{
filter_bold_specimens(bold_specimen_records, taxon, mapping_file_name)
}
\arguments{
\item{bold_specimen_records}{a data.frame of BOLD specimen records obtained with bold::bold_specimens}

\item{mapping_file_name}{Path to a mapping file previously created with check_and_do_mapping}
}
\description{
This function works on a BOLD specimen record (tsv format)
obtained using the bold::bold_specimens function.
It then filters to retain only those specimens not yet present in the mapping file
(since these records were already obtained from NCBI). The BOLD IDs as listed in NCBI
seem to map incompletely to the processids however, so an extra filtering step
is also performed, to remove any records that have the text 'NCBI' in the institution_storing field.
It returns the ******
}
\examples{
filter_bold_specimens(bold_specimens(taxon="Athericidae", format="tsv"), mapping_file_name="./Athericidae_mapping.txt")
}
\keyword{bold}
\keyword{metabarcoding}
