% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_taxonomic_levels.R
\name{format_taxonomic_levels}
\alias{format_taxonomic_levels}
\title{Format a classification data frame into dada2 reference database FASTA header format}
\usage{
format_taxonomic_levels(classif)
}
\arguments{
\item{classif}{classification data frame for an individual taxon, created by taxize::classification}
}
\description{
This is a function to take a classification data frame (in the format
returned by the 'taxize::classification' function) and return the taxonomy
information formatted as a string of the format required by dada2 for its
reference database. For example,
Metazoa;Arthropoda;Insecta;Diptera;Chironomidae;Chironomus;Chironomus riparius
}
\examples{
format_taxonomic_levels(classif=Athericidae_classification[[1]])
}
\keyword{classification}
\keyword{dada2}
\keyword{ncbi}
\keyword{taxize}
\keyword{taxonomy}
\keyword{uid}
