% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_extraction_functions.R
\name{get_accession_version_ids_from_fasta_file}
\alias{get_accession_version_ids_from_fasta_file}
\title{get_accession_version_ids_from_fasta_file}
\usage{
get_accession_version_ids_from_fasta_file(raw_fasta_file_name)
}
\arguments{
\item{raw_fasta_file_name}{Path to the fasta file}
}
\description{
This function extracts NCBI accession.version ID from a NCBI-retrieved FASTA file
It returns the extracted IDs only, as a character vector
}
\examples{
get_accession_version_ids_from_fasta_file(raw_fasta_file_name="Chlorophyta.fasta")
}
\keyword{fasta}
\keyword{metabarcoding}
\keyword{ncbi}
\keyword{sequence}
