% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_extraction_functions.R
\name{get_existing_bold_from_mapping_file}
\alias{get_existing_bold_from_mapping_file}
\title{get_existing_bold_from_mapping_file}
\usage{
get_existing_bold_from_mapping_file(mapping_file_name)
}
\arguments{
\item{mapping_file_name}{Path to the mapping file}
}
\description{
This function extracts BOLD IDs from a mapping file created by the *** function
It returns the extracted IDs only, as a character vector
}
\examples{
get_existing_bold_from_mapping_file(mapping_file_name="Chlorophyta_mapping.txt")
}
\keyword{BOLD}
\keyword{id}
\keyword{taxonomy}
