% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_from_id_vector.R
\name{get_fasta_from_id_vector}
\alias{get_fasta_from_id_vector}
\title{Get NCBI FASTA from Unique Identifiers}
\usage{
get_fasta_from_id_vector(id_vector, block_size = 5000)
}
\arguments{
\item{id_vector}{A character vector of NCBI UIDs.}

\item{block_size}{An integer giving the size of blocks into which the full vector of IDs with missing classification info is split. This is important for ensuring full retrieval in the case of taxa with a large number of records (more than approx. 10000). Default = 5000}
}
\description{
This function retrieves FASTA sequences for a given vector of NCBI unique identifiers (UIDs)
This function is quite slow because it doesn't use the web history functionality.
It returns the fasta file in Entrez API fasta format.
}
\examples{
get_fasta_from_id_vector(id_vector=c('1072897717', '329757399', '259881131'), taxon="Hypogasturidae", block_size=5000)
}
\keyword{fasta}
\keyword{metabarcoding}
\keyword{ncbi}
\keyword{uid}
