% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_with_web_history.R
\name{get_fasta_with_web_history}
\alias{get_fasta_with_web_history}
\title{get_fasta_with_web_history}
\usage{
get_fasta_with_web_history(ncbi_search_block)
}
\arguments{
\item{ncbi_search_block}{An entry in a list created with the get_ncbi_search_blocks function. Contains a vector of NCBI UIDs ('block_ids'), the web history cookie ('cookie') and query key ('qk'), and the total number of records found with the underlying query.}
}
\description{
This function retrieves FASTA sequences for a given ncbi_block (created with ***)
This function is quite fast because it uses the web history present in the ncbi_block.
It returns the fasta file in Entrez API fasta format.
}
\examples{
get_fasta_with_web_history(ncbi_search_block=ncbi_search_blocks[[1]], taxon="Chlorophyta", block_size=5000)
}
\keyword{fasta}
\keyword{metabarcoding}
\keyword{ncbi}
\keyword{uid}
