% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_naming_functions.R
\name{make_classification_name}
\alias{make_classification_name}
\title{make_classification_name}
\usage{
make_classification_name(taxon, directory, suffix = "_classification.rds")
}
\arguments{
\item{taxon}{taxon name (character string)}

\item{directory}{path to the directory containing the classification file (character string)}

\item{suffix}{desired suffix for the file. Defaults to "_classification.rds"}
}
\description{
This function creates the file path for the classification file (an RDS object)
}
\examples{
make_mapping_name(taxon="Chlorophyta", directory="~/Documents/metabarcoding/ref_datafiles/", suffix="_classification.rds")
}
\keyword{filename}
