% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_naming_functions.R
\name{make_dada2_fasta_name}
\alias{make_dada2_fasta_name}
\title{make_dada2_fasta_name}
\usage{
make_dada2_fasta_name(taxon, directory, suffix = "_dada2.fasta")
}
\arguments{
\item{taxon}{taxon name (character string)}

\item{directory}{path to the directory containing the fasta file (character string)}

\item{suffix}{desired suffix for the file. Defaults to "_dada2.fasta"}
}
\description{
This function creates the file path for the dada2-formatted fasta file (fasta format)
}
\examples{
make_mapping_name(taxon="Chlorophyta", directory="~/Documents/metabarcoding/ref_datafiles/", suffix="_dada2.fasta")
}
\keyword{filename}
