% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_naming_functions.R
\name{make_mapping_name}
\alias{make_mapping_name}
\title{make_mapping_name}
\usage{
make_mapping_name(taxon, directory, suffix = "_mapping.txt")
}
\arguments{
\item{taxon}{taxon name (character string)}

\item{directory}{path to the directory containing the mapping file (character string)}

\item{suffix}{desired suffix for the file. Defaults to "_mapping.txt"}
}
\description{
This function creates the file path for the mapping file (data.frame)
}
\examples{
make_mapping_name(taxon="Chlorophyta", directory="~/Documents/metabarcoding/ref_datafiles/", suffix="_mapping.txt")
}
\keyword{filename}
