% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_uids_to_bold_and_accn.R
\name{map_uids_to_bold_and_accn}
\alias{map_uids_to_bold_and_accn}
\title{Map NCBI Unique Identifiers (UIDs) to Accession.Version and BOLD IDs}
\usage{
map_uids_to_bold_and_accn(id_vector, taxon, chunk_size = 60)
}
\arguments{
\item{id_vector}{A character vector of NCBI UIDs}

\item{taxon}{Taxon name as a character string.}

\item{chunk_size}{An integer giving the size of chunks into which id_vector will be split for efficient extraction. Defaults to 60.}
}
\description{
This function extracts the BOLD (Barcode of Life Database) IDs, if present, and the accession.version IDs
from NCBI for a potentially very large input character vector of NCBI UIDs.
It returns a dataframe with columns uid, bold, and accn.
If any errors are encountered (e.g. in connecting to the Entrez API), the relevant chunk will be skipped
and the extraction will continue.
}
\examples{
map_uids_to_bold_and_accn(id_vector=c('1072897717', '329757399', '259881131'), taxon="Hypogasturidae", chunk_size=60)
}
\keyword{bold}
\keyword{id}
\keyword{ncbi}
