% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_taxonomic_levels.R
\name{reduce_taxonomic_levels}
\alias{reduce_taxonomic_levels}
\title{Reduce taxonomic levels in a classification data frame}
\usage{
reduce_taxonomic_levels(classif, taxonomic_levels)
}
\arguments{
\item{classif}{classification data frame for an individual taxon, created by taxize::classification}

\item{taxonomic_levels}{character vector of the names of the taxonomic levels to retain}
}
\description{
This is a function to take a classification data frame (in the format
returned by the 'taxize::classification' function) and return new data frame
containing only the classification for the taxonomic levels specified
in the character vector 'taxonomic_levels'
}
\examples{
reduce_taxonomic_levels(classif=Athericidae_classification[[1]], taxonomic_levels=c("kingdom", "phylum", "class", "order", "family", "genus", "species"))
}
\keyword{classification}
\keyword{ncbi}
\keyword{taxize}
\keyword{taxonomy}
\keyword{uid}
